Add Warning For /locate for Admin+
Hello,

In the real Mineplex, when a staff member /locates a Admin+, they get a notification telling them that the member located them, this is for privacy purposes of the high-ups. If you ask a mod to locate an admin or something like that they will tell you that they will get in trouble. Additionally, I made the /locate command look much nicer, I worked really hard to make sure this works and it is completely tested and bug-free.

Thanks
Techno3600